Update copyright year to 2024 and holder to Wilhelm Payne
The changes in the `LICENSE.txt` file involve updating the copyright information. Specifically, the year has been updated to 2024, and the name of the copyright holder has been changed to Wilhelm Payne. This ensures that the license reflects the current year and the correct individual or entity holding the copyright.